Fix unexpected alb.Listener target group refresh (#4083)
Fixes unexpected alb.Listener target group refresh.

The change was fixed upstream but patched out due to an unrelated
problem causing a panic in Pulumi. The root cause is now fixed in the
bridge. This change drops the two obsolete patches to catch up with
upstream, and enrolls alb.Listener in PlanResourceChange version of the
bridge that fixes the issue.

Fixes #3427
Fixes #3722 
Fixes #3723
Fixes #4079