Fix ListenerRule panic if no forward parameter is missing in defaultActions
In #4083 we removed two patches fixing panics for Listener and
ListenerRule resources and instead enrolled the resources in
PlanResourceChange.

The ListenerRule (lb.ListenerRule and alb.ListenerRule) resources
were missed to be enrolled in PlanResourceChange, which caused
the original panic to surface again.

This change enrolls the ListenerRule resources in PlanResourceChange
and adds a regression test covering those resources.