chore: TestIMDSAuth is very slow
The `TestIMDSAuth` test frequently takes ~45 minutes to run. Looking
into it more it looks like the `command:remote:CopyFile` takes the bulk
of the time and I am not sure why. Copying the file to S3 via
`BucketObjectV2` and then using the userData to copy the file from S3
during instance start up seems to be significantly faster