chore: SkipInstall for tests that do not need it
Some of the YAML tests only need the AWS provider. Unless the SkipInstall option is specified, they will attempt to
download the latest version of the provider before realizing there is an ambient in-process version to test with. This
can cause sporadic CI failures unrelated to the test. Specifying SkipInstall optimizes away this redundant step.

Example failure log addressed here:

    pulumiTest.go:103: failed to install packages and plugins: exit status 255
          resource plugin aws installing
          warning: Error downloading plugin: unknown version for plugin aws
          Will retry in 80ms [1/5]
          warning: Error downloading plugin: unknown version for plugin aws
          Will retry in 160ms [2/5]