Disable experimental post-quantum key exchange mechanism X25519Kyber768Draft00
The AWS Provider was upgraded to Go 1.23 in v6.51.0, which introduced a change
to the crypto/tls standard library package. It enabled the post-quantum
key exchange mechanism `X25519Kyber768Draft00` by default. This experimental key
exchange mechanism is causing errors in the AWS firewall.
As a short term workaround this change disables the experimental key exchange mechanism.

Upstream maintainers and AWS are in touch to work on a long-term fix.