Re-generate schema to fix aws.iam.Role description (#4589)
Fixes https://github.com/pulumi/pulumi-aws/issues/4588

This simply a result of executing `make tfgen`. There is no impact on
SDKs its seems. There should be however impact on the registry.

This is a little un-intuitive but schema.json does not get reused for
SDK generation. Instead we have something like this:

```
schemaJson := tfgen("schema")
nodeSDK := tfgen("node")
javaSDK := tfgen("java")
```

It looks like CI is missing a check that re-generating schema.json with
`make tfgen` is a no-change operation, allowing this to drift and break
registry. Filing an issue in ci-mgmt.

https://github.com/pulumi/ci-mgmt/issues/1090