Discard log messages when initializing the provider
Messages sent early to the "log" package end up being forwarded to stderr, and Pulumi displays stderr by default to
users, which is undesirable.

Later in the provider life-cycle, Pulumi Terraform Bridge will reset `log.SetOutput` again to capture and
appropriately track diagnostic logs, which should still work with this change.

Fixes #4645

A test is added to ensure we do not leak diagnostics again on a simple program.