Fix IMDSv2 test to allow /tmp to be immutable (#4673)
The IMDSv2 test provisions a RC of the pulumi-aws-provider inside an AWS
test account and checks if it can authenticate. Unfortunately something
changed in the cloud that no longer permits unix programs to make
changes to `/tmp`. The test relied to this for no good reason; the
updated version creates a dedicated folder /home/ec2-user/repro to
perform the check.

Fixes https://github.com/pulumi/pulumi-aws/issues/4670