Update ci-mgmt to strip debug symbols from provider binaries (#4786)
This change pulls the latest ci-mgmt changes in order to reduce the size
of provider binaries. In detail, we're now setting the `-s -w` linker
flags to strip debug symbols. This cuts the provider binary size by
roughly 300MB, resulting in a 32% smaller provider binary.

Note: I also had to change `toolVersions.node` to `toolVersions.nodejs`
because the configuration is now type checked. The `toolVersions.node`
configuration option was renamed to `nodejs` in this PR:
https://github.com/pulumi/ci-mgmt/pull/1063