chore: remove minimal schema from source (#4829)
Keeping a copy of minimal schema (introduced in
https://github.com/pulumi/pulumi-aws/pull/4587) is proving to conflict
with Pulumi tooling such as `upgrade-provider` and is not strictly
necessary. With this change the minimal schema itself as well as CI
checks to make sure it is up to date are removed. Instead it will be
computed on-the-fly by the release job as before:
 
- GitHub actions invoke this:
https://github.com/pulumi/pulumi-aws/blob/916e6f28039407d7dcb1a295c6f9e43613dd6f87/.github/workflows/build_provider.yml#L60

- Make extension ensures that this runs go generate prior to the actual
build so that the minimal schema is re-computed:
https://github.com/pulumi/pulumi-aws/blob/916e6f28039407d7dcb1a295c6f9e43613dd6f87/.mk/minimal_schema.mk#L9

Fixes:

- https://github.com/pulumi/pulumi-aws/issues/4811
- https://github.com/pulumi/pulumi-aws/issues/4702