Remove unused npm dependencies
It looks like `resolve` and `builtin-modules` dependencies aren't actually used anywhere, so stop depending on them.

The motivation for this change is that the dependency on `resolve` causes errors to be logged in logs during plugin discovery, because the package includes its tests that have malformed `package.json` files in its package. If we don't actually need the dependency, it'd be nice to remove it so those errors are no longer logged (so users don't have to ask us about them).

Also remove the `builtin-modules` dependency while cleaning this up, since it doesn't appear to be used either AFAICT.