[7.0.0-alpha]: remove tags related patches (#5505)
There are two main breaking changes in this update.

1. If resource tags are removed _and_ the keys that are removed also
exist in `DefaultTags` _with a value of ""_, then they will not be
updated
related to
[#37297](https://github.com/hashicorp/terraform-provider-aws/issues/37297)
- Created a follow up issue
https://github.com/pulumi/pulumi-aws/issues/5517
2. `tags` and `tagsAll` are now separate things. Previously we set
`tags`=`tagsAll` so `tags` always included _all_ the tags. Now only
`tagsAll` will contain all tags.

I kept some of our customization on the Pulumi side to deal with a
Pulumi only issue. For sdkv2 resources, if you are removing
`defaultTags` then the `tagsAll` attribute will only be updated when:
   - `pulumi refresh --run-program` with the changes to defaultTags
   - `pulumi up` to propagate the changes