[7.0.0-alpha]: upgrade upstream to 6.0.0-beta (#5479)
This PR performs a major version upgrade while upgrading upstream to the
`release/6.0.0-beta` branch.

Some things that occur as part of this major version upgrade are:
- Resources/properties that were deprecated in v6 are removed
- Changes to `MaxItemsOne` that were held off in v6 occur now

In additional the following groups of patches were removed. This was
included in this PR rather than in a follow up because as part of the
`6.0.0-beta` upstream completely removed aws sdkv1. In order to keep
these patches around we would have had to either include it back or
update the patches to use aws sdkv2.

- [Remove sdkv1 related patches](#5466)
closes #5461, closes #5460, closes #5465, closes #5463, closes #5470,
closes #5472

- [Legacy S3 Bucket Patches](#5473)
closes #3720, closes #3726, closes #2727, closes #2728, closes #5474,
closes #2752, closes #2729

- [Remove deprecated Gamelift MatchMaking resources](#4375)
closes #5409, closes #2745, closes #2742, closes #2740, closes #5467,
closes #5468, closes #5469

There are a lot of changes in this PR that I won't be able to enumerate
in this description so I have created a follow up PR that includes a doc
with all the details. https://github.com/pulumi/pulumi-aws/pull/5498

closes #2732