Remove Region overlay to support removing dangling ref for Region (#5508)
**Update 2**

Looking at the potential risk of customers losing `provider.Region` if
set via the environment, and the time tradeoffs implementing
string-Enums on provider types via engine and codegen, the team suggests
moving forward with removing _some_ Node functionality so we are able to
remove the dangling reference.

Specifically:

- Implementing enums-that-are-strings on providers via engine and
codegen presents a week's worth of work if done by a non-Core
contributor plus waiting on releases.
- We would be able to keep the `getEnvOrDefault` utilities in each
provider - Region is a string.
- All SDKs behave the same.
- We can keep strongly typed enums - they do exist on every SDK; only
when used as provider inputs/outputs they must be converted to strings.
For example:
```
func main() {
	pulumi.Run(func(ctx *pulumi.Context) error {
		providerRegion := aws.RegionUSEast2

		provider, err := aws.NewProvider(ctx, "prov", &aws.ProviderArgs{
			Region: pulumi.String(providerRegion),
		})
		if err != nil {
			return err
		}

		ctx.Export("pRegion", provider.Region)

		return nil
	})
}
```
is what is done today.

The changes presented in this pull request would result in Node
developers having to cast their Region enums as strings before passing
to the provider, as in the Go example above. Additionally, Node would
lose the strongly typed Output in favor of a string. (I'm not clear on
this making any practical difference whatsoever - if Region fields take
strings, and the Region Output is now a string, that should just
continue to work.)

This pull request presents that approach, for the reasons above.

**End Update 2**


**Update**

There's a path forward to allow us to keep strongly typed Enums.

I filed https://github.com/pulumi/pulumi/issues/19583 to track that p/p
should allow enums that are strings in provider configs.
I opened https://github.com/pulumi/pulumi/pull/19585 to implement.
The fix currently has the drawback that `defaultInfo` breaks on these
fields, because the utilities code doesn't yet know how to properly
unwrap the new strong types.
However, I think this is an acceptable tradeoff: I filed
https://github.com/pulumi/pulumi/issues/19584 as a followup,
and while we would lose the `defaultInfo` on the Region field until that
is addressed, the provider should still be able to pick up environment
variables from the environment.
[Sample
diff](https://github.com/pulumi/pulumi-aws/compare/674ec9959e24ea0c0131b33c32a021f221206370...guin/remove-region-thursday).

**End Update**


This change aligns the Node SDK with the other providers.

Prior to this change, the schema has `"$ref":
"#/types/aws:index/region:Region"` in the definition for
`config.Region`. This allows the Node SDK to use an overlay that defines
a Region type.

However, this ref does not match the Region enum type defined in
`types.go`. While all SDKs have explicit Regions defined in `enums` as a
result of `ExtraTypes`, the provider config for Go, Dotnet, etc, is
currently only implemented as a basic string (probably because the type
reference is invalid here and thus not used in codegen).

By removing the (invalid for non-Node SDKs) type token ref , and
removing the overlay file, this will result in config.Region being a
string in the Node SDK as well.

It appears as though the Enum capability [does not currently work on
provider config vars](https://github.com/pulumi/pulumi/issues/13435),
but I could be misreading the situation. This would mean that the enum
types are not being used at all by the provider type.

Alternatives attempted:

- Changing the type ref to "$ref": "#/types/aws:index/Region:Region" -
this results in the loss of the Output for Region in the Node SDK
- Changing the type in types.go to be "aws:index/region:Region" - this
has the same effect as uppercasing the type ref - the tokens now match
and codegen doesn't handle the config type ref well:
```
/pulumi/pulumi-aws/sdk/dotnet/Provider.cs(397,22): error CS0029: Cannot implicitly convert type 'string' to 'Pulumi.Input<Pulumi.Aws.Region>'
```
- AltTypes does not resolve the above issue either.

I'm open to more suggestions in case I'm missing something here. It's
obviously not ideal to move the Node SDK from a stronger-typed Region
enum to a string, but it appears as though enums aren't supported
completely in the provider config, and the only way to get strong types
is to have an overlay; but having that overlay results in a dangling
schema type.