Remove several dangling resource refs from schema (#5549)
In support of removing dangling schema references, this pull request
converts the following resource refs, which were implemented via
AltTypes, to using _only_ the corresponding string refs of either `name`
or `id` field:

- "$ref": "#/types/aws:ec2/launchConfiguration:LaunchConfiguration"
- "$ref": "#/types/aws:ec2/placementGroup:PlacementGroup"
- "$ref": "#/types/aws:cloudwatch/logGroup:LogGroup"
- "$ref": "#/types/aws:sns/topic:Topic"
- "$ref": "#/types/aws:elasticbeanstalk/application:Application"

Reference the updates to the migration guide to see which code changes
will need to be made.
Four new integration tests were added to show that there are no diffs on
these fields during a provider upgrade and applying the code change. It
is a code-only change.

Not resolved:

- "$ref": "#/types/aws:apigateway/restApi:RestApi"
- "$ref": "#/types/aws:apigateway/deployment:Deployment"
- "$ref": "#/types/aws:iam/instanceProfile:InstanceProfile"
- "$ref": "#/types/aws:iam/group:Group"
- "$ref": "#/types/aws:iam/role:Role"
- "$ref": "#/types/aws:iam/user:User"
- "$ref": "#/types/aws:iot/policy:Policy"
- "$ref": "#/types/aws:lambda/function:Function"
- "$ref":
"#/types/aws:elasticbeanstalk/applicationVersion:ApplicationVersion"

Part of #5540.