Migrate applicationversion to applicationversion.name in v7 (#5555)
This pull request resolves the final dangling reference to what is a
resource type.

Instead of `version: appVersion`, users will now need to use `version:
appVersion.name`. This is documented, and the case is added to the test.

edit: also a test for lambdaFunction is enabled that was missed
previously.

Closes https://github.com/pulumi/pulumi-aws/issues/5540.