Add PolicyDocument type to schema
This fully schematizes the `PolicyDocument` type and replaces the node
only overlay definition. I've added an upgrade test to ensure that the
types are the same and won't cause type errors for users in v7.

The benefit with this approach is that we get to keep the
`PolicyDocument` type _and_ make it available to the other languages.

There are a couple of open issues that we have to workaround.

1. https://github.com/pulumi/pulumi/issues/8324
Go gen will generate unused imports. To workaround this we have to add
the `PolicyDocument` type to the individual modules that don't already
import from `iam`.

2. https://github.com/pulumi/pulumi-terraform-bridge/issues/3103
The `Condition` type doesn't generate in Go so we are using the `any`
type for now

fixes #5338