Pin CI to pulumi v3.187.0 (#5732)
Still investigating root cause, but seems to be a performance regression
in the Pulumi CLI. GitHub support confirms that:

> Regarding the shutdown signal I observed in the logs, it typically
indicates that the runner's resources were exhausted during the job,
which prevented communication between Actions and the runner from being
maintained.

Testing on `3.187.0` confirms this is the last stable version. We need
to unblock CI, we can switch back once pulumi has been fixed.

re #5731, re #5726