Remove skipping examples gen for recursive resources (#5727)
This pull request reverts the skipping of examples generation for the
following three resources:

- `wafv2:RuleGroup`
- `wafv2:Acl`
- `appsync:GraphQLApi`

It appears from the typical duration for the `make schema` (~30min) and
`make build_go` (~30min) CI runs on this pull request, that the
conditions for #2598 have been met.

NOTE: current CI timeouts are unrelated to the changes.

This pull request closes:

- #5691
- #2598

- **Unskip examples generation and make schema for 3 resources**
- **make SDKs**