Remove rds.db.Proxy revert patch
This patch was added a while ago to fix a panic in the bridge. We
accidentally broke it [a couple of releases
ago](https://github.com/pulumi/pulumi-aws/pull/5872/files#diff-621a91eb03679b13105bec390bcc509e1358fd96caff45b2437709e3d526c775).

While fixing the patch, I went ahead and tested what would happen if we
removed the patch altogether and it worked.

It looks like the reason we didn't catch this regression in tests is
that our test only ran in preview. This must have been good enough to
catch the original panic, but wasn't good enough to catch this
regression. I've updated the tests to perform actual deployments and
added another test case that was mentioned in the original issue.

closes #3968
fixes #5901